optimize(parsercore): Shrink parser-core cell and boundary structs
- Store the classification transaction of a ClassifiedBoundary as a uint32 value. The field now uses 4 bytes instead of 8 bytes.
- Return an error from currentClassificationTransaction when the transaction exceeds the uint32 range. ClassifyBoundary and ClassifyBoundaryWithRow check that error before they build a boundary.
- Compare the stored boundary transaction as a uint64 value in validateClassification, so recorded transactions still match.
- Store headerIndex and selectedOrdinal of the generic scheduler cell as int32 values. Call sites convert them back to int at the slice and API edges.
- Add a test that ClassifyBoundary rejects a transaction larger than math.MaxUint32.
- These structs are copied for every dispatch pass and every boundary. The narrower fields lower bytes per pass and cap the growth of the compact cell slices.

Buckley-Change-Hash: sha256:4c601040eb09556b65965a6d5c916990f55153517bed6998e3caaa735daaf1bf
Buckley-Change-Stats: files=4 insertions=66 deletions=44 binaries=0

+++ parsercore_phase0_classified_internal_test.go
@@ -15,7 +15,7 @@ func mustDiagnosticParserCoreGenericCell(t testing.TB, compact *core.Core, heade
 	if err != nil {
 		t.Fatal(err)
 	}
-	return diagnosticParserCoreGenericCell{headerIndex: headerIndex, boundary: boundary}
+	return diagnosticParserCoreGenericCell{headerIndex: int32(headerIndex), boundary: boundary}
 }
 
 func TestDiagnosticParserCoreClassifiedBoundaryAndReductionPlanShape(t *testing.T) {
